remove no-unused-import rule
while moving tests from no-unused-import to no-unused-var
