+++ docs/rules.md
@@ -17,7 +17,6 @@ title:       "Rule Index of Solhint"
 | [no-console](./rules/best-practises/no-console.md)                       | No console.log/logInt/logBytesX/logString/etc & No hardhat and forge-std console.sol import statements                                     | ✔️          |
 | [no-empty-blocks](./rules/best-practises/no-empty-blocks.md)             | Code block has zero statements inside. Some common exceptions apply.                                                                       | ✔️          |
 | [no-global-import](./rules/best-practises/no-global-import.md)           | Import statement includes an entire file instead of selected symbols                                                                       | ✔️          |
-| [no-unused-import](./rules/best-practises/no-unused-import.md)           | Reports a warning when an imported name is not used                                                                                        | ✔️          |
 | [no-unused-vars](./rules/best-practises/no-unused-vars.md)               | Variable "name" is unused.                                                                                                                 | ✔️          |
 | [one-contract-per-file](./rules/best-practises/one-contract-per-file.md) | Enforces the use of ONE Contract per file see [here](https://docs.soliditylang.org/en/v0.8.21/style-guide.html#contract-and-library-names) | ✔️          |
 | [payable-fallback](./rules/best-practises/payable-fallback.md)           | When fallback is not payable you will not be able to receive ethers.                                                                       | ✔️          |
