+++ lib/rules/best-practises/index.js
@@ -9,12 +9,11 @@ const PayableFallbackChecker = require('./payable-fallback')
 const ReasonStringChecker = require('./reason-string')
 const NoConsoleLogChecker = require('./no-console')
 const NoGlobalImportsChecker = require('./no-global-import')
-const NoUnusedImportsChecker = require('./no-unused-import')
 const CustomErrorsChecker = require('./custom-errors')
 const ExplicitTypesChecker = require('./explicit-types')
 const OneContractPerFileChecker = require('./one-contract-per-file')
 
-module.exports = function checkers(reporter, config, inputSrc, tokens) {
+module.exports = function checkers(reporter, config, inputSrc) {
   return [
     new CodeComplexityChecker(reporter, config),
     new InterfaceStartsWithIChecker(reporter),
@@ -27,7 +26,6 @@ module.exports = function checkers(reporter, config, inputSrc, tokens) {
     new ReasonStringChecker(reporter, config),
     new NoConsoleLogChecker(reporter),
     new NoGlobalImportsChecker(reporter),
-    new NoUnusedImportsChecker(reporter, tokens),
     new CustomErrorsChecker(reporter),
     new ExplicitTypesChecker(reporter),
     new OneContractPerFileChecker(reporter),
